fix: コマンドの表示判定

+++ src/main/kotlin/me/notsmatch/kyoshubot/command/AddCommand.kt
@@ -23,7 +23,11 @@ class AddCommand(val boshuService: BoshuService,  val settingsService: GuildSett
     override fun execute(event: CommandEvent?) {
         event?.apply {
 
-            event.message.delete().complete()
+            val settings = settingsService.getGuildSettings(guild.idLong)
+
+            if (settings.getCommandOption("add") == null || !settings.getCommandOption("add")!!.visibility) {
+                event.message.delete().complete()
+            }
 
             val boshu = boshuService.getBoshu(guild.idLong, channel.idLong)
                 ?: return replyInDm(EmbedBuilder().apply {
@@ -132,8 +136,6 @@ class AddCommand(val boshuService: BoshuService,  val settingsService: GuildSett
                         addField("タイトル", title, true)
                     }.build()
                 )*/
-
-                val settings = settingsService.getGuildSettings(guild.idLong)
 
                 textChannel.editMessageById(boshu.messageId, EmbedBuilder().apply {
                     setColor(Color.CYAN)
